Add aarch64, llvm, static target
We are using this target and it seems to be working fine so it would be
nice to have it on cachix. However when I tried to run the "Test
{default,example} shell" they failed.

The only failing crate is rockdb:

/nix/store/...-glibc-2.37-8-dev/include/gnu/stubs.h:7:11: fatal error:
'gnu/stubs-32.h' file not found

The other crates compile fine. This is on x86_64 NixOS.